refactor(auth): isolate console sanitizer
