+++ src/agents/auth-profiles/state-observation.ts
@@ -1,6 +1,6 @@
 import { redactIdentifier } from "../../logging/redact-identifier.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
-import { sanitizeForConsole } from "../pi-embedded-error-observation.js";
+import { sanitizeForConsole } from "../console-sanitize.js";
 import type { AuthProfileFailureReason, ProfileUsageStats } from "./types.js";
 
 const observationLog = createSubsystemLogger("agent/embedded");
