+++ src/agents/pi-embedded-error-observation.ts
@@ -1,8 +1,11 @@
 import { readLoggingConfig } from "../logging/config.js";
 import { redactIdentifier } from "../logging/redact-identifier.js";
 import { getDefaultRedactPatterns, redactSensitiveText } from "../logging/redact.js";
+import { sanitizeForConsole } from "./console-sanitize.js";
 import { getApiErrorPayloadFingerprint, parseApiErrorInfo } from "./pi-embedded-helpers.js";
 import { stableStringify } from "./stable-stringify.js";
+
+export { sanitizeForConsole } from "./console-sanitize.js";
 
 const MAX_OBSERVATION_INPUT_CHARS = 64_000;
 const MAX_FINGERPRINT_MESSAGE_CHARS = 8_000;
@@ -39,30 +42,6 @@ function boundObservationInput(text: string | undefined): string | undefined {
   return trimmed.length > MAX_OBSERVATION_INPUT_CHARS
     ? trimmed.slice(0, MAX_OBSERVATION_INPUT_CHARS)
     : trimmed;
-}
-
-export function sanitizeForConsole(text: string | undefined, maxChars = 200): string | undefined {
-  const trimmed = text?.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  const withoutControlChars = Array.from(trimmed)
-    .filter((char) => {
-      const code = char.charCodeAt(0);
-      return !(
-        code <= 0x08 ||
-        code === 0x0b ||
-        code === 0x0c ||
-        (code >= 0x0e && code <= 0x1f) ||
-        code === 0x7f
-      );
-    })
-    .join("");
-  const sanitized = withoutControlChars
-    .replace(/[\r\n\t]+/g, " ")
-    .replace(/\s+/g, " ")
-    .trim();
-  return sanitized.length > maxChars ? `${sanitized.slice(0, maxChars)}…` : sanitized;
 }
 
 function replaceRequestIdPreview(
